User - 주문 화면 결제기능 제외 완료.

+++ src/main/java/hello/aiofirstuser/domain/OrderItem.java
@@ -23,5 +23,5 @@ public class OrderItem {
     @JoinColumn(name = "product_variant_id")
     private ProductVariant productVariant;
 
-    private Long quantity;
+    private int quantity;
 }
